Move module operations into its own file
